Suppress hidden files in ArchiveImportService; imiplement Caesar cipher; misc

+++ source/java/org/rsna/ctp/pipeline/AbstractImportService.java
@@ -122,44 +122,46 @@ public abstract class AbstractImportService extends AbstractPipelineStage implem
 		File file;
 		if (queueManager != null) {
 			while ((file = queueManager.dequeue(active)) != null) {
-				lastFileOut = file;
-				lastTimeOut = System.currentTimeMillis();
-				FileObject fileObject = FileObject.getInstance(lastFileOut);
-				fileObject.setStandardExtension();
+				if (file.length() > 0) {
+					lastFileOut = file;
+					lastTimeOut = System.currentTimeMillis();
+					FileObject fileObject = FileObject.getInstance(lastFileOut);
+					fileObject.setStandardExtension();
 
-				if (logDuplicates) {
-					//*********************************************************************************************
-					//See if this object has the same UID as a recent one.
-					String currentUID = fileObject.getUID();
-					if (recentUIDs.contains(currentUID)) {
-						logger.warn("----------------------------------------------------------------");
-						logger.warn(name);
-						logger.warn("Duplicate UID in last "+maxQueueSize+" objects: "+currentUID);
-						String s = "";
-						long time = 0;
-						for (int i=0; i<recentUIDs.size(); i++) {
-							String uid = recentUIDs.get(i);
-							s += uid.equals(currentUID) ? "!" : "*";
-							time = recentTimes.get(i).longValue();
+					if (logDuplicates) {
+						//*********************************************************************************************
+						//See if this object has the same UID as a recent one.
+						String currentUID = fileObject.getUID();
+						if (recentUIDs.contains(currentUID)) {
+							logger.warn("----------------------------------------------------------------");
+							logger.warn(name);
+							logger.warn("Duplicate UID in last "+maxQueueSize+" objects: "+currentUID);
+							String s = "";
+							long time = 0;
+							for (int i=0; i<recentUIDs.size(); i++) {
+								String uid = recentUIDs.get(i);
+								s += uid.equals(currentUID) ? "!" : "*";
+								time = recentTimes.get(i).longValue();
+							}
+							long deltaT = System.currentTimeMillis() - time;
+							logger.warn("[oldest] "+s+"! [newest]  deltaT = "+deltaT+"ms");
+							logger.warn("----------------------------------------------------------------");
 						}
-						long deltaT = System.currentTimeMillis() - time;
-						logger.warn("[oldest] "+s+"! [newest]  deltaT = "+deltaT+"ms");
-						logger.warn("----------------------------------------------------------------");
+						recentUIDs.add(currentUID);
+						recentTimes.add( new Long( System.currentTimeMillis() ) );
+						if (recentUIDs.size() > maxQueueSize) { recentUIDs.remove(); recentTimes.remove(); }
+						//*********************************************************************************************
 					}
-					recentUIDs.add(currentUID);
-					recentTimes.add( new Long( System.currentTimeMillis() ) );
-					if (recentUIDs.size() > maxQueueSize) { recentUIDs.remove(); recentTimes.remove(); }
-					//*********************************************************************************************
+
+					//Make sure we accept objects of this type.
+					if (acceptable(fileObject)) return fileObject;
+
+					//If we get here, this import service does not accept
+					//objects of the active type. Try to quarantine the
+					//object, and if that fails, delete it.
+					if (quarantine != null)  quarantine.insert(fileObject);
+					else fileObject.getFile().delete();
 				}
-
-				//Make sure we accept objects of this type.
-				if (acceptable(fileObject)) return fileObject;
-
-				//If we get here, this import service does not accept
-				//objects of the active type. Try to quarantine the
-				//object, and if that fails, delete it.
-				if (quarantine != null)  quarantine.insert(fileObject);
-				else fileObject.getFile().delete();
 			}
 		}
 		return null;
